requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
passlib[bcrypt]
PyJWT
python-dotenv
requests
httpx
